requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Bioinformatics
lifelines>=0.27.0

# Utilities
requests>=2.31.0
pyyaml>=6.0

# Jupyter (optional, for exploration)
jupyter>=1.0.0
jupyterlab>=4.0.0

# For static Plotly exports (optional)
kaleido>=0.2.0